+++ R/get_c14data.R
@@ -143,7 +143,8 @@ get_all_parser_functions <- function() {
     "pacea" = c14bazAAR::get_pacea,
     "14cpalaeolithic" = c14bazAAR::get_14cpalaeolithic,
     "medafricarbon" = c14bazAAR::get_medafricarbon,
-    "jomon" = c14bazAAR::get_jomon
+    "jomon" = c14bazAAR::get_jomon,
+    "katsianis" = c14bazAAR::get_katsianis
   ))
 }